backend/requirements.txt
aiohttp==3.8.5
aiopg==1.3.5
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
boto3==1.24.78
botocore==1.27.78
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-postgres==0.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
debugpy==1.6.3
decorator==5.1.1
Django==4.0
django-appconf==1.0.5
django-core==1.4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.1
django-jsonfield==1.4.1
django-model-utils==4.3.1
django-notifications-hq==1.8.2
django-notify-x==0.1.9
django-phone-field==1.8.1
django-select2==8.1.1
django-sslserver==0.22
django-storages==1.13.1
django-user-sessions==2.0.0
django-xforwardedfor-middleware==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
entrypoints==0.4
executing==1.0.0
frozenlist==1.4.0
geoip2==4.7.0
google-crc32c==1.5.0
gunicorn==20.1.0
httplib2==0.20.4
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
jmespath==1.0.1
jsonfield==3.1.0
jupyter-core==4.11.1
jupyter_client==7.3.5
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
maxminddb==2.4.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.5
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==4.21.5
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
Pygments==2.13.0
PyJWT==2.4.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
swapper==1.3.0
tk==0.1.0
tornado==6.2
traitlets==5.3.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
watchdog==3.0.0
wcwidth==0.2.5
Werkzeug==2.2.3
yarl==1.9.2
zope.interface==6.0
